requirements.txt
pyboy
pillow
openai
